fix warning about bitwise or between u32 and size_t
(Ported from Mellanox OFED 2.4)


Orabug: 27510513

(cherry picked from commit ffc4093edecd754a98cccbbe7b9eb57595762d58)
cherry-pick-repo=linux-uek.git

+++ drivers/infiniband/hw/mlx4/qp.c
@@ -3466,7 +3466,7 @@ static void build_tunnel_header(struct ib_ud_wr *wr, void *wqe, unsigned *mlx_se
 	if (sizeof (hdr) <= spc) {
 		memcpy(inl + 1, &hdr, sizeof (hdr));
 		wmb();
-		inl->byte_count = cpu_to_be32(1 << 31 | sizeof (hdr));
+		inl->byte_count = cpu_to_be32(1 << 31 | (u32)sizeof (hdr));
 		i = 1;
 	} else {
 		memcpy(inl + 1, &hdr, spc);
